requirements.txt
spaces
torch
torchvision
ultralytics
Pillow